requirements.txt
beautifulsoup4==4.12.3
flask==3.0.3
openai==1.68.2
python-dotenv==1.0.1
requests==2.32.3
